fix deployment

+++ website/deploy.sh
@@ -7,6 +7,20 @@
 # fix: git push origin master --force
 #      where master is the branch name
 
+if [[ "$#" -eq 0 ]]; then
+	echo "You must give an OS"
+	echo "./deploy.sh windows   OR   ./deploy.sh macos"
+	exit 1
+elif [[ "$1" == "windows" || "$1" == "macos" ]]; then
+	OS = "$1"
+else
+	echo "Unrecognized OS:" "$1"
+	exit 1
+fi
+
+# OS="windows"
+# OS="macos"
+
 # repo and branch to save website too
 REPO="github.com/MarsUniversity/ece387"
 BRANCH="gh-pages"
@@ -22,10 +36,14 @@ if [[ ${TRAVIS} ]]; then
 	git config user.name ${USER}
 	REMOTE="https://${GITHUB_TOKEN}@${REPO}"
 # setup for local manual push
-else
+elif [[ ${OS} == "windows" ]]; then
+	TMP="/c/Users/Kevin.Walchko/tmp/tmpdir"
+	# REMOTE="git@github.com:MarsUniversity/ece387.git"
+	REMOTE="https://github.com/MarsUniversity/ece387.git"
+	rm -fr ${TMP}
+elif [[ ${OS} == "macos" ]]; then
 	TMP="/var/tmp/tmpdir"
 	REMOTE="git@github.com:MarsUniversity/ece387.git"
-	rm -fr ${TMP}
 fi
 
 # website data is stored here:
